feat(form-geo): Move form to client component
